refactor(inquiry): decomposes aspects of affiliated view; organizes hooks
