+++ src/v2/Apps/Debug/DebugInquiryApp.tsx
@@ -1,7 +1,7 @@
 // TODO: Remove once inquiry is complete
 import { Button } from "@artsy/palette"
 import React from "react"
-import { useInquiry } from "v2/Components/Inquiry/useInquiry"
+import { useInquiry } from "v2/Components/Inquiry/Hooks/useInquiry"
 
 export const DebugInquiryApp: React.FC<{}> = () => {
   const { inquiryQuestionnaire, showInquiry } = useInquiry({
