+++ src/v2/Components/Inquiry/__tests__/views/InquiryCommercialInterest.jest.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 import { mount } from "enzyme"
 import { InquiryCommercialInterest } from "../../views/InquiryCommercialInterest"
-import { useUpdateMyUserProfile } from "../../useUpdateMyUserProfile"
+import { useUpdateMyUserProfile } from "../../Hooks/useUpdateMyUserProfile"
 import { useInquiryContext } from "../../InquiryContext"
 import { flushPromiseQueue } from "v2/DevTools"
