+++ src/v2/Components/Inquiry/__tests__/views/InquiryLogin.jest.tsx
@@ -2,7 +2,7 @@ import { mount } from "enzyme"
 import React from "react"
 import { InquiryLogin } from "../../views/InquiryLogin"
 import { login } from "v2/Utils/auth"
-import { useArtworkInquiryRequest } from "../../useArtworkInquiryRequest"
+import { useArtworkInquiryRequest } from "../../Hooks/useArtworkInquiryRequest"
 import { flushPromiseQueue } from "v2/DevTools"
 import { useInquiryContext } from "../../InquiryContext"
 import { fill } from "../util"
