+++ src/v2/Components/Inquiry/__tests__/views/InquirySignUp.jest.tsx
@@ -2,7 +2,7 @@ import { mount } from "enzyme"
 import React from "react"
 import { flushPromiseQueue } from "v2/DevTools"
 import { InquirySignUp } from "../../views/InquirySignUp"
-import { useArtworkInquiryRequest } from "../../useArtworkInquiryRequest"
+import { useArtworkInquiryRequest } from "../../Hooks/useArtworkInquiryRequest"
 import { signUp } from "v2/Utils/auth"
 import { useInquiryContext } from "../../InquiryContext"
 import { fill } from "../util"
