+++ src/v2/Components/Inquiry/views/InquiryLogin.tsx
@@ -12,7 +12,7 @@ import React, { useState } from "react"
 import { createRelaySSREnvironment } from "v2/System/Relay/createRelaySSREnvironment"
 import { wait } from "v2/Utils/wait"
 import { useInquiryContext } from "../InquiryContext"
-import { useArtworkInquiryRequest } from "../useArtworkInquiryRequest"
+import { useArtworkInquiryRequest } from "../Hooks/useArtworkInquiryRequest"
 import { login } from "v2/Utils/auth"
 import { useInquiryAccountContext, Screen } from "./InquiryAccount"
 import { logger } from "../util"
